Move PixelFormat and Rect below rfb.py so the imports point one way
rfb.py sat at the top of the dependency graph -- it imports the decoders
package -- and also defined PixelFormat and Rect, so everything below it
that needed those types had to import back up. pixelformat.py imported
rfb purely to name rfb.PixelFormat, and each decoder module carried a
TYPE_CHECKING block with string-literal annotations to avoid a runtime
cycle.

PixelFormat moves to pixelformat.py, which now imports nothing else from
the package. Rect moves to decoders/base.py, beside the decodeForClient
signature that receives it: it is the pump-to-decoder contract type, and
the pump is the only thing that constructs one. rfb.py imports both and
re-exports them, since rfb.PixelFormat and rfb.Rect are public API --
client.py, loggingproxy.py, the tests and docs/modules.rst all reach for
them there, and rfb.PixelFormat is the same object as before.

Nothing pickles or isinstance-checks PixelFormat and the goldens store
bytes, so the class's new module-qualified name breaks nothing. Fields,
STRUCT and behaviour are untouched.

docs/modules.rst gains a pixelformat section: automodule documents only
members a module defines, so without it PixelFormat would have dropped
out of the published docs when it stopped being defined in rfb.py.

The one remaining TYPE_CHECKING guard, RectBuffer in decoders/base.py,
is not the same problem -- buffer.py raises DecodeError and so imports
base.py, which is a cycle within one layer rather than an inverted one.
