+++ vncdotool/decoders/copyrect.py
@@ -2,21 +2,18 @@
 from __future__ import annotations
 
 from struct import unpack
-from typing import TYPE_CHECKING, ClassVar, Iterator
+from typing import ClassVar, Iterator
 
 from ..const import Encoding
-from .base import ClientDecoder
-
-# rfb.py imports this package, so importing from it at runtime is a cycle.
-if TYPE_CHECKING:  # pragma: no cover
-    from ..rfb import PixelFormat, Rect
+from ..pixelformat import PixelFormat
+from .base import ClientDecoder, Rect
 
 
 class CopyRectDecoder(ClientDecoder):
     ENCODING: ClassVar[Encoding] = Encoding.COPY_RECTANGLE
 
     def decodeForClient(
-        self, client: object, rect: "Rect", pixel_format: "PixelFormat"
+        self, client: object, rect: Rect, pixel_format: PixelFormat
     ) -> Iterator[int]:
         block = yield 4
         srcx, srcy = unpack("!HH", block)
